Output/primary/src/arraylist.rs

pub type ArrayListValue = *mut ();

pub struct ArrayList {
    pub data: Vec<ArrayListValue>,
    pub length: usize,
    pub _alloced: usize,
}

pub type ArrayListEqualFunc = fn(ArrayListValue, ArrayListValue) -> i32;
pub type ArrayListCompareFunc = fn(ArrayListValue, ArrayListValue) -> i32;
pub fn arraylist_new(length: usize) -> Option<Box<ArrayList>> {
    let mut length = length;

    if length == 0 {
        length = 16;
    }

    let new_arraylist = Box::new(ArrayList {
        _alloced: length,
        length: 0,
        data: Vec::with_capacity(length),
    });

    Some(new_arraylist)
}
pub fn arraylist_free(arraylist: &mut Option<ArrayList>) {
    if let Some(al) = arraylist.take() {
        drop(al.data);
    }
}
// Modified: Removed duplicate definition of `ArrayListValue`
// and implemented the `Default` trait manually for `ArrayListValue`
struct ArrayListValue {
    // Example fields
    value: i32,
}

// Modified: Manually implemented the `Default` trait for `ArrayListValue`
impl Default for ArrayListValue {
    fn default() -> Self {
        ArrayListValue { value: 0 }
    }
}

// Modified: Removed duplicate definition of `ArrayList`
struct ArrayList {
    _alloced: usize,
    data: Vec<ArrayListValue>,
}

pub fn arraylist_enlarge(arraylist: &mut ArrayList) -> bool {
    let newsize = arraylist._alloced * 2;

    if newsize > arraylist.data.len() {
        // Modified: Ensure `ArrayListValue` implements `Default` trait
        arraylist.data.resize(newsize, ArrayListValue::default());
    }

    arraylist._alloced = newsize;
    true
}